+++ go.mod
@@ -1,10 +1,8 @@
 module github.com/djthorpe/sensors
 
 require (
-	github.com/abiosoft/ishell v2.0.0+incompatible
 	github.com/djthorpe/gopi v1.0.20
 	github.com/djthorpe/gopi-hw v1.0.5
 	github.com/djthorpe/mutablehome v0.0.0-20180502123221-577c352e5cba
-	github.com/flynn-archive/go-shlex v0.0.0-20150515145356-3f9db97f8568 // indirect
-	github.com/olekukonko/tablewriter v0.0.0-20180506121414-d4647c9c7a84
+	github.com/olekukonko/tablewriter v0.0.0-20180912035003-be2c049b30cc
 )
