+++ protocol/openthings_test.go
@@ -12,11 +12,11 @@ import (
 	_ "github.com/djthorpe/sensors/protocol/openthings"
 )
 
-func OTTest_000(t *testing.T) {
+func Test_OT_000(t *testing.T) {
 	// Create an OOK module
 	if app, err := gopi.NewAppInstance(gopi.NewAppConfig("sensors/protocol/openthings")); err != nil {
 		t.Fatal(err)
-	} else if _, ok := app.ModuleInstance("sensors/protocol/openthings").(sensors.ProtoOT); ok == false {
-		t.Fatal("OOK does not comply to ProtoOT interface")
+	} else if _, ok := app.ModuleInstance("sensors/protocol/openthings").(sensors.OTProto); ok == false {
+		t.Fatal("OOK does not comply to OTProto interface")
 	}
 }
